Condense all fields to only Person class and delete Tutor and Student class
